+++ src/client/model.ts
@@ -9,6 +9,9 @@ import { Memory, Blit, GetBlits, GetMemoryAfterDma, Copper } from './dma';
 import { DisplayUnit, scaleValue } from './display';
 import { DmaRecord } from '../backend/profile_types';
 
+import { Base64Decoder } from 'base64-encoding';
+let base64decoder: Base64Decoder;
+
 /**
  * Category of call frames. Grouped into system, modules, and user code.
  */
@@ -209,7 +212,7 @@ const ensureSourceLocations = (profile: ICpuProfileRaw): readonly IAnnotationLoc
 /**
  * Computes the model for the given profile.
  */
-export const buildModel = (profile: ICpuProfileRaw): IProfileModel => {
+export const buildModel = async (profile: ICpuProfileRaw): Promise<IProfileModel> => {
 	if(!profile.timeDeltas)
 		profile.timeDeltas = [];
 	if(!profile.samples)
@@ -228,6 +231,9 @@ export const buildModel = (profile: ICpuProfileRaw): IProfileModel => {
 	}
 */
 	console.time('buildModel');
+
+	if(!base64decoder)
+		base64decoder = await new Base64Decoder().optimize();
 
 	const sourceLocations = ensureSourceLocations(profile);
 	const locations: ILocation[] = sourceLocations.map((l, id) => {
@@ -338,8 +344,8 @@ export const buildModel = (profile: ICpuProfileRaw): IProfileModel => {
 
 		// decode memory to binary
 		if(model.amiga.chipMem) {
-			const chipMem = Uint8Array.from(atob(model.amiga.chipMem), (c) => c.charCodeAt(0));
-			const bogoMem = Uint8Array.from(atob(model.amiga.bogoMem), (c) => c.charCodeAt(0));
+			const chipMem = base64decoder.decode(model.amiga.chipMem);
+			const bogoMem = base64decoder.decode(model.amiga.bogoMem);
 			model.memory = new Memory(chipMem, bogoMem);
 		}
 	}