+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to the "amiga-debug" extension will be documented in this fi
 - CHG: screen: zoom panel size, clipping
 - FIX: fix memory content for multi-frame captures (was wrong after frame 1)
 - NEW: screen: show bitplane pointers
+- CHG: screen: improve performance by 40%
+- CHG: profiler: improve loading performance by >1000%
 
 ## 1.5.6
 - FIX: disassembly: fix divul, divsl, FPU, fmoveml (see `68k-dis.ts`)